requirements.txt
azure-functions
azure-storage-blob
azure-ai-formrecognizer
openai